# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Horde Portal Bot API
  version: 1.0.0
  description: API REST pour bots et agents CRM
servers:
  - url: https://portal.hordeagence.com/api/v1
    description: Production
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: "API Key au format: horde_xxxxx"
  schemas:
...
    ClientStatus:
      type: string
      enum: [lead, contacted, in_project, pending_review, completed, archived]
    Contact:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        role:
          $ref: '#/components/schemas/ContactRole'
        is_primary:
          type: boolean
        notes:
          type: string
    ContactRole:
      type: string
      enum: [decision_maker, technical, billing, marketing, other]
...
    Stats:
      type: object
      properties:
        counts:
          type: object
          properties:
            clients:
              type: integer
            pipeline:
              type: integer
            projects:
              type: integer
            followups:
              type: integer
        clients_by_status:
          type: array
          items:
            type: object
            properties:
              status:
                $ref: '#/components/schemas/ClientStatus'
              count:
                type: integer
        projects_by_status:
          type: array
          items:
            type: object
            properties:
              status:
                type: string
              count:
                type: integer
        metrics:
          type: object
          properties:
            conversion_rate:
              type: integer
            messages_30d:
              type: integer
            new_clients_30d:
              type: integer
    Error:
      type: object
      properties:
        error:
          type: string
paths:
...
  /contacts/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    patch:
      summary: Update contact
      tags: [Contacts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                phone:
                  type: string
                role:
                  $ref: '#/components/schemas/ContactRole'
                is_primary:
                  type: boolean
                notes:
                  type: string
      responses:
        '200':
          description: Contact updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Contact'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Delete contact
      tags: [Contacts]
      responses:
        '200':
          description: Contact deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
  /stats:
    get:
      summary: Get dashboard statistics
      tags: [Stats]
      responses:
        '200':
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stats'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'